+++ R/tm_t_pp_laboratory.R
@@ -7,7 +7,6 @@
 #' @param timepoints (`character`)\cr name of time variable used for
 #' the laboratory table.
 #' @param anrind (`character`)\cr name of the analysis reference range indicator variable.
-#' @param avalu (`character`)\cr name of the analysis value unit variable.
 #' @param patient_id (`character`)\cr patient ID.
 #' @param round_value (`numeric`)\cr number of decimal places to be used when rounding.
 #' @keywords internal
@@ -19,7 +18,8 @@ template_laboratory <- function(dataname = "ANL",
                                 timepoints = "ADY",
                                 aval = lifecycle::deprecated(),
                                 aval_var = "AVAL",
-                                avalu = "AVALU",
+                                avalu = lifecycle::deprecated(),
+                                avalu_var = "AVALU",
                                 patient_id = NULL,
                                 round_value = 0L) {
   if (lifecycle::is_present(aval)) {
@@ -31,6 +31,15 @@ template_laboratory <- function(dataname = "ANL",
     )
   }
 
+  if (lifecycle::is_present(avalu)) {
+    avalu_var <- avalu
+    warning(
+      "The `avalu` argument of `template_laboratory()` is deprecated as of teal.modules.clinical 0.8.16. ",
+      "Please use the `avalu_var` argument instead.",
+      call. = FALSE
+    )
+  }
+
   assertthat::assert_that(
     assertthat::is.string(dataname),
     assertthat::is.string(paramcd),
@@ -38,7 +47,7 @@ template_laboratory <- function(dataname = "ANL",
     assertthat::is.string(anrind),
     assertthat::is.string(timepoints),
     assertthat::is.string(aval_var),
-    assertthat::is.string(avalu),
+    assertthat::is.string(avalu_var),
     is.integer(round_value) && round_value >= 0
   )
 
@@ -51,7 +60,7 @@ template_laboratory <- function(dataname = "ANL",
       expr = {
         dataname[, aval_char] <- round(dataname[, aval_char], round_value)
         labor_table_base <- dataname %>%
-          dplyr::select(timepoints, paramcd, param, aval_var, avalu, anrind) %>%
+          dplyr::select(timepoints, paramcd, param, aval_var, avalu_var, anrind) %>%
           dplyr::arrange(timepoints) %>%
           dplyr::select(-timepoints) %>%
           dplyr::group_by(paramcd, param) %>%
@@ -91,7 +100,7 @@ template_laboratory <- function(dataname = "ANL",
         paramcd = as.name(paramcd),
         aval_var = as.name(aval_var),
         aval_char = aval_var,
-        avalu = as.name(avalu),
+        avalu_var = as.name(avalu_var),
         timepoints = as.name(timepoints),
         anrind = as.name(anrind),
         patient_id = patient_id,
@@ -120,7 +129,7 @@ template_laboratory <- function(dataname = "ANL",
 #' \code{ANRIND} column of the `ADLB` dataset with 3 possible levels "HIGH", "LOW" and "NORMAL".
 #' @param aval_var ([teal.transform::choices_selected()] or [teal.transform::data_extract_spec()])\cr
 #' \code{AVAL} column of the `ADLB` dataset.
-#' @param avalu ([teal.transform::choices_selected()] or [teal.transform::data_extract_spec()])\cr
+#' @param avalu_var ([teal.transform::choices_selected()] or [teal.transform::data_extract_spec()])\cr
 #' \code{AVALU} column of the `ADLB` dataset.
 #' @inheritParams module_arguments
 #'
@@ -164,7 +173,7 @@ template_laboratory <- function(dataname = "ANL",
 #'         choices = variable_choices(ADLB, "AVAL"),
 #'         selected = "AVAL"
 #'       ),
-#'       avalu = choices_selected(
+#'       avalu_var = choices_selected(
 #'         choices = variable_choices(ADLB, "AVALU"),
 #'         selected = "AVALU"
 #'       )
@@ -182,7 +191,8 @@ tm_t_pp_laboratory <- function(label,
                                timepoints = NULL,
                                aval = lifecycle::deprecated(),
                                aval_var = NULL,
-                               avalu = NULL,
+                               avalu = lifecycle::deprecated(),
+                               avalu_var = NULL,
                                param = NULL,
                                paramcd = NULL,
                                anrind = NULL,
@@ -199,6 +209,17 @@ tm_t_pp_laboratory <- function(label,
     aval <- aval_var # resolves missing argument error
   }
 
+  if (lifecycle::is_present(avalu)) {
+    avalu_var <- avalu
+    warning(
+      "The `avalu` argument of `tm_t_pp_laboratory()` is deprecated as of teal.modules.clinical 0.8.16. ",
+      "Please use the `avalu_var` argument instead.",
+      call. = FALSE
+    )
+  } else {
+    avalu <- avalu_var # resolves missing argument error
+  }
+
   logger::log_info("Initializing tm_t_pp_laboratory")
   checkmate::assert_string(label)
   checkmate::assert_string(dataname)
@@ -211,7 +232,7 @@ tm_t_pp_laboratory <- function(label,
   data_extract_list <- list(
     timepoints = `if`(is.null(timepoints), NULL, cs_to_des_select(timepoints, dataname = dataname)),
     aval_var = `if`(is.null(aval_var), NULL, cs_to_des_select(aval_var, dataname = dataname)),
-    avalu = `if`(is.null(avalu), NULL, cs_to_des_select(avalu, dataname = dataname)),
+    avalu_var = `if`(is.null(avalu_var), NULL, cs_to_des_select(avalu_var, dataname = dataname)),
     param = `if`(is.null(param), NULL, cs_to_des_select(param, dataname = dataname)),
     paramcd = `if`(is.null(paramcd), NULL, cs_to_des_select(paramcd, dataname = dataname)),
     anrind = `if`(is.null(anrind), NULL, cs_to_des_select(anrind, dataname = dataname))
@@ -240,7 +261,7 @@ ui_g_laboratory <- function(id, ...) {
   is_single_dataset_value <- teal.transform::is_single_dataset(
     ui_args$timepoints,
     ui_args$aval_var,
-    ui_args$avalu,
+    ui_args$avalu_var,
     ui_args$param,
     ui_args$paramcd,
     ui_args$anrind
@@ -257,7 +278,7 @@ ui_g_laboratory <- function(id, ...) {
       teal.reporter::simple_reporter_ui(ns("simple_reporter")),
       ###
       shiny::tags$label("Encodings", class = "text-primary"),
-      teal.transform::datanames_input(ui_args[c("timepoints", "aval_var", "avalu", "param", "paramcd", "anrind")]),
+      teal.transform::datanames_input(ui_args[c("timepoints", "aval_var", "avalu_var", "param", "paramcd", "anrind")]),
       teal.widgets::optionalSelectInput(
         ns("patient_id"),
         "Select Patient:",
@@ -289,9 +310,9 @@ ui_g_laboratory <- function(id, ...) {
         is_single_dataset = is_single_dataset_value
       ),
       teal.transform::data_extract_ui(
-        id = ns("avalu"),
+        id = ns("avalu_var"),
         label = "Select AVALU variable:",
-        data_extract_spec = ui_args$avalu,
+        data_extract_spec = ui_args$avalu_var,
         is_single_dataset = is_single_dataset_value
       ),
       teal.transform::data_extract_ui(
@@ -324,7 +345,7 @@ srv_g_laboratory <- function(id,
                              patient_col,
                              timepoints,
                              aval_var,
-                             avalu,
+                             avalu_var,
                              param,
                              paramcd,
                              anrind,
@@ -379,7 +400,7 @@ srv_g_laboratory <- function(id,
       data_extract = list(
         timepoints = timepoints,
         aval_var = aval_var,
-        avalu = avalu,
+        avalu_var = avalu_var,
         param = param,
         paramcd = paramcd,
         anrind = anrind
@@ -388,7 +409,7 @@ srv_g_laboratory <- function(id,
       select_validation_rule = list(
         timepoints = shinyvalidate::sv_required("Please select timepoints variable."),
         aval_var = shinyvalidate::sv_required("Please select AVAL variable."),
-        avalu = shinyvalidate::sv_required("Please select AVALU variable."),
+        avalu_var = shinyvalidate::sv_required("Please select AVALU variable."),
         param = shinyvalidate::sv_required("Please select PARAM variable."),
         paramcd = shinyvalidate::sv_required("Please select PARAMCD variable."),
         anrind = shinyvalidate::sv_required("Please select ANRIND variable.")
@@ -418,7 +439,7 @@ srv_g_laboratory <- function(id,
         dataname = "ANL",
         timepoints = input[[extract_input("timepoints", dataname)]],
         aval_var = input[[extract_input("aval_var", dataname)]],
-        avalu = input[[extract_input("avalu", dataname)]],
+        avalu_var = input[[extract_input("avalu_var", dataname)]],
         param = input[[extract_input("param", dataname)]],
         paramcd = input[[extract_input("paramcd", dataname)]],
         anrind = input[[extract_input("anrind", dataname)]],